support asynchronous input
