+++ viewer/src/pages/RoutingMain.tsx
@@ -1,4 +1,4 @@
-import React from "react";
+import {Suspense} from "react";
 import { Routes, Route, Navigate } from "react-router-dom";
 
 import ModelListPage from "pages/list";
@@ -7,12 +7,14 @@ import ModelViewPage from "pages/view"
 export const RoutingMain = () => {
     console.log("route called")
     return (
-        <Routes>
-            <Route path="/models/list" element={<ModelListPage />} />
-            <Route path="/models/item/:modelId/view/:lod" element={<ModelViewPage />} />
-            <Route path="/models" element={<Navigate to="/models/list" replace />} />
-            <Route path="*" element={<Navigate to="/models" replace />} />
-        </Routes>
+        <Suspense fallback={null}>
+            <Routes>
+                <Route path="/models/list" element={<ModelListPage />} />
+                <Route path="/models/item/:modelId/view/:lod" element={<ModelViewPage />} />
+                <Route path="/models" element={<Navigate to="/models/list" replace />} />
+                <Route path="*" element={<Navigate to="/models" replace />} />
+            </Routes>
+        </Suspense>
     )
 }
