+++ viewer/src/styles.css
@@ -1,5 +1,5 @@
 .detail-window {
-    height: 250px;
+    height: 270px;
     width: 270px;
     position: relative;
     padding-top: 10px;
